Restructurization - DashboardCarousel nested in Dasboard, helper and tempData moved to folder helpers
