+++ src/screens/dashboard/Dashboard.tsx
@@ -4,19 +4,34 @@ import { Box, Text } from 'utils/theme'
 import { ScrollView } from 'react-native'
 import { SafeAreaWrapper } from 'components/SafeAreaWrapper'
 import { DashboardHeader } from 'screens/dashboard/components/DashboardHeader'
-import { DashboardCarousel } from 'screens/dashboard/components/DashboardCarousel'
-import { ValidationOfGroupDayOff, USER_GROUPS_DAYS_OFF } from 'screens/dashboard/temporaryData'
+import {
+  ValidationOfGroupDayOff,
+  USER_GROUPS_DAYS_OFF,
+} from 'screens/dashboard/helpers/temporaryData'
 import { TeamElement } from 'screens/dashboard/components/TeamElement'
+import { CarouselElement } from 'screens/dashboard/components/CarouselElement'
+import { dataToBeDisplayed, ValidationOfDataToBeDisplayed } from 'screens/dashboard/helpers/helper'
 
 export const Dashboard: FC = () => {
-  const { t } = useTranslation('dashboard')
+  const { t, i18n } = useTranslation('dashboard')
   const teamsList: ValidationOfGroupDayOff[] = USER_GROUPS_DAYS_OFF
+  const companyHolidaysData: ValidationOfDataToBeDisplayed[] = dataToBeDisplayed(i18n.language)
 
   return (
     <SafeAreaWrapper isDefaultBgColor>
       <DashboardHeader />
       <ScrollView>
-        <DashboardCarousel />
+        <ScrollView horizontal showsHorizontalScrollIndicator={false}>
+          {companyHolidaysData.map((item) => (
+            <CarouselElement
+              key={item.id}
+              isOnHoliday={item.isOnHoliday}
+              firstName={item.user.firstName}
+              lastName={item.user.lastName}
+              dayToBeDisplayed={item.dayToBeDisplayed}
+            />
+          ))}
+        </ScrollView>
         <Box marginTop="l">
           <Text variant="header" marginHorizontal="m">
             {t('teamsList').toUpperCase()}
